+++ pkg/ingester/ingester_ring_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/go-kit/log"
 	"github.com/grafana/dskit/flagext"
 	"github.com/grafana/dskit/ring"
 	"github.com/stretchr/testify/assert"
@@ -28,8 +29,9 @@ func TestRingConfig_DefaultConfigToLifecyclerConfig(t *testing.T) {
 	expected.FinalSleep = cfg.FinalSleep
 	expected.ReadinessCheckRingHealth = false
 	expected.HeartbeatPeriod = 15 * time.Second
+	expected.RingTokenGenerator = ring.NewRandomTokenGenerator()
 
-	assert.Equal(t, expected, cfg.ToLifecyclerConfig())
+	assert.Equal(t, expected, cfg.ToLifecyclerConfig(log.NewNopLogger()))
 }
 
 func TestRingConfig_CustomConfigToLifecyclerConfig(t *testing.T) {
@@ -44,16 +46,18 @@ func TestRingConfig_CustomConfigToLifecyclerConfig(t *testing.T) {
 	cfg.ExcludedZones = []string{"zone-a", "zone-b"}
 	cfg.TokensFilePath = "/tokens.file"
 	cfg.NumTokens = 1234
-	cfg.InstanceID = "instance"
+	cfg.InstanceID = "instance-10"
 	cfg.InstanceInterfaceNames = []string{"abc"}
 	cfg.InstancePort = 1111
 	cfg.InstanceAddr = "1.2.3.4"
-	cfg.InstanceZone = "zone-X"
+	cfg.InstanceZone = "zone-c"
 	cfg.UnregisterOnShutdown = true
 	cfg.ObservePeriod = 10 * time.Minute
 	cfg.MinReadyDuration = 3 * time.Minute
 	cfg.FinalSleep = 2 * time.Minute
 	cfg.ListenPort = 10
+	cfg.TokenGeneratorStrategy = spreadMinimizingTokenGeneration
+	cfg.SpreadMinimizingZones = []string{"zone-a", "zone-b", "zone-c"}
 
 	// The lifecycler config should be generated based upon the ingester ring config
 	expected.RingConfig.KVStore.Store = "memberlist"
@@ -85,5 +89,67 @@ func TestRingConfig_CustomConfigToLifecyclerConfig(t *testing.T) {
 	expected.Addr = cfg.InstanceAddr
 	expected.ListenPort = cfg.ListenPort
 
-	assert.Equal(t, expected, cfg.ToLifecyclerConfig())
+	logger := log.NewNopLogger()
+	tokenGenerator, err := ring.NewSpreadMinimizingTokenGenerator(expected.ID, expected.Zone, cfg.SpreadMinimizingZones, logger)
+	assert.NoError(t, err)
+	expected.RingTokenGenerator = tokenGenerator
+
+	assert.Equal(t, expected, cfg.ToLifecyclerConfig(logger))
+}
+
+func TestRingConfig_CustomTokenGenerator(t *testing.T) {
+	instanceID := "instance-10"
+	instanceZone := "zone-a"
+	wrongZone := "zone-d"
+	spreadMinimizingZones := []string{"zone-a", "zone-b", "zone-c"}
+
+	tests := map[string]struct {
+		zone                    string
+		tokenGenerationStrategy string
+		spreadMinimizingZones   []string
+		expectedResultStrategy  string
+	}{
+		"spread-min-tokens and correct zones give a SpreadMinimizingTokenGenerator": {
+			zone:                    instanceZone,
+			tokenGenerationStrategy: spreadMinimizingTokenGeneration,
+			spreadMinimizingZones:   spreadMinimizingZones,
+			expectedResultStrategy:  spreadMinimizingTokenGeneration,
+		},
+		"spread-min-tokens and a wrong zone give a RandomTokenGenerator": {
+			zone:                    wrongZone,
+			tokenGenerationStrategy: spreadMinimizingTokenGeneration,
+			spreadMinimizingZones:   spreadMinimizingZones,
+			expectedResultStrategy:  randomTokenGeneration,
+		},
+		"random-tokens gives a RandomTokenGenerator": {
+			zone:                    instanceZone,
+			tokenGenerationStrategy: randomTokenGeneration,
+			expectedResultStrategy:  randomTokenGeneration,
+		},
+		"unknown token generation strategy gives a RandomTokenGenerator": {
+			zone:                    instanceZone,
+			tokenGenerationStrategy: "bla-bla-tokens",
+			expectedResultStrategy:  randomTokenGeneration,
+		},
+	}
+
+	for _, testData := range tests {
+		cfg := RingConfig{}
+		cfg.InstanceID = instanceID
+		cfg.InstanceZone = testData.zone
+		cfg.TokenGeneratorStrategy = testData.tokenGenerationStrategy
+		cfg.SpreadMinimizingZones = testData.spreadMinimizingZones
+		lifecyclerConfig := cfg.ToLifecyclerConfig(log.NewNopLogger())
+		if testData.expectedResultStrategy == randomTokenGeneration {
+			tokenGenerator, ok := lifecyclerConfig.RingTokenGenerator.(*ring.RandomTokenGenerator)
+			assert.True(t, ok)
+			assert.NotNil(t, tokenGenerator)
+		} else if testData.expectedResultStrategy == spreadMinimizingTokenGeneration {
+			tokenGenerator, ok := lifecyclerConfig.RingTokenGenerator.(*ring.SpreadMinimizingTokenGenerator)
+			assert.True(t, ok)
+			assert.NotNil(t, tokenGenerator)
+		} else {
+			assert.Fail(t, "This case is not supported")
+		}
+	}
 }